requirements.txt
django>4.0.0,<4.1.0
djangorestframework
django-cors-headers
djoser
django-versatileimagefield

